+++ changelog.md
@@ -11,4 +11,5 @@ Notable changes of Reviewer Recommender after version 3.0.4 will be documented i
   If the email address got from AMiner is illegal, it won't be filled. And an alert will be displayed.
 
 - Fix the bug of failing to analyze keywords. (analyze.js)
-- fix the bug of failing to create author list. (analyze.js)
+- Fix the bug of failing to create author list. (analyze.js)
+- Add function to add current reviewers and authors to the reviewer roster for this extension.(analyze.js)